SoundManager의 이름을 SoundSystem으로 변경한다
싱글톤..이름의 일관성..System...
